+++ package.json
@@ -11,6 +11,7 @@
     "start": "npm run dev"
   },
   "dependencies": {
+    "@ant-design/icons": "^5.2.6",
     "@types/styled-components": "^5.1.32",
     "antd": "^5.11.5",
     "react": "^18.2.0",
